+++ packages/seedless-onboarding-controller/src/SeedlessOnboardingController.ts
@@ -23,7 +23,6 @@ import { secp256k1 } from '@noble/curves/secp256k1';
 import { Mutex } from 'async-mutex';
 
 import {
-  assertIsAuthUserInfoValid,
   assertIsPasswordOutdatedCacheValid,
   assertIsSeedlessOnboardingUserAuthenticated,
   assertIsValidVaultData,
@@ -46,7 +45,6 @@ import type {
   SeedlessOnboardingControllerState,
   AuthenticatedUserDetails,
   SocialBackupsMetadata,
-  SRPBackedUpUserDetails,
   VaultEncryptor,
   RefreshJWTToken,
   RevokeRefreshToken,
@@ -375,7 +373,7 @@ export class SeedlessOnboardingController<
     groupedAuthConnectionId?: string;
     socialLoginEmail?: string;
     refreshToken: string;
-    revokeToken: string;
+    revokeToken?: string;
     skipLock?: boolean;
   }) {
     const doAuthenticateWithNodes = async () => {
@@ -409,8 +407,10 @@ export class SeedlessOnboardingController<
           state.socialLoginEmail = socialLoginEmail;
           state.metadataAccessToken = metadataAccessToken;
           state.refreshToken = refreshToken;
-          // Temporarily store revoke token & access token in state for later vault creation
-          state.revokeToken = revokeToken;
+          if (revokeToken) {
+            // Temporarily store revoke token & access token in state for later vault creation
+            state.revokeToken = revokeToken;
+          }
           state.accessToken = accessToken;
 
           // we will check if the controller state is properly set with the authenticated user info
@@ -902,7 +902,7 @@ export class SeedlessOnboardingController<
         }
       }
 
-      assertIsAuthUserInfoValid(this.state);
+      this.#assertIsAuthenticatedUser(this.state);
       const {
         nodeAuthTokens,
         authConnectionId,
@@ -953,20 +953,21 @@ export class SeedlessOnboardingController<
   /**
    * Check if the user is authenticated with the seedless onboarding flow by checking the token values in the state.
    *
+   * This method will check the `accessToken` and `revokeToken` in the state, besides the social login authentication details.
+   * If both are present, the user is authenticated.
+   * If either is missing, the user is not authenticated.
+   *
+   * This method is useful when we want to check if the state has valid authenticated user details to perform vault creations.
+   *
    * @returns True if the user is authenticated, false otherwise.
    */
-  async checkIsSeedlessOnboardingUserAuthenticated(): Promise<boolean> {
-    let isAuthenticated = false;
+  async getIsUserAuthenticated(): Promise<boolean> {
     try {
-      assertIsSeedlessOnboardingUserAuthenticated(this.state);
-      isAuthenticated = true;
+      this.#assertIsAuthenticatedUser(this.state);
+      return Boolean(this.state.accessToken) && Boolean(this.state.revokeToken);
     } catch {
-      isAuthenticated = false;
+      return false;
     }
-    this.update((state) => {
-      state.isSeedlessOnboardingUserAuthenticated = isAuthenticated;
-    });
-    return isAuthenticated;
   }
 
   #setUnlocked(): void {
@@ -1110,8 +1111,12 @@ export class SeedlessOnboardingController<
    * @returns The authentication public key.
    */
   #recoverAuthPubKey(): SEC1EncodedPublicKey {
-    this.#assertIsSRPBackedUpUser(this.state);
     const { authPubKey } = this.state;
+    if (!authPubKey) {
+      throw new Error(
+        SeedlessOnboardingControllerErrorMessage.SRPNotBackedUpError,
+      );
+    }
 
     return base64ToBytes(authPubKey);
   }
@@ -1126,7 +1131,7 @@ export class SeedlessOnboardingController<
   async #recoverEncKey(
     password: string,
   ): Promise<Omit<RecoverEncryptionKeyResult, 'rateLimitResetResult'>> {
-    assertIsAuthUserInfoValid(this.state);
+    this.#assertIsAuthenticatedUser(this.state);
     const {
       nodeAuthTokens,
       authConnectionId,
@@ -1539,7 +1544,8 @@ export class SeedlessOnboardingController<
   }): Promise<void> {
     this.#assertIsAuthenticatedUser(this.state);
 
-    const { revokeToken, accessToken } = this.state;
+    const { accessToken, revokeToken } =
+      await this.#getAccessTokenAndRevokeToken(password);
 
     const vaultData: DeserializedVaultData = {
       toprfAuthKeyPair: rawToprfAuthKeyPair,
@@ -1609,6 +1615,48 @@ export class SeedlessOnboardingController<
         state.encryptedSeedlessEncryptionKey = bytesToBase64(encryptedKey);
       });
     });
+  }
+
+  /**
+   * Get the access token and revoke token from the state or the vault.
+   *
+   * @param password - The password to decrypt the vault.
+   * @returns The access token and revoke token.
+   */
+  async #getAccessTokenAndRevokeToken(
+    password: string,
+  ): Promise<{ accessToken: string; revokeToken: string }> {
+    let { accessToken, revokeToken } = this.state;
+    // `accessToken` and `revokeToken` are both available in the state, `ONLY` when the wallet (vault) is unlocked
+    // or during the period between the social authentication and the vault creation during the onboarding flow.
+    if (accessToken && revokeToken) {
+      return { accessToken, revokeToken };
+    }
+
+    // if `password` is provided to decrypt the vault, decrypt the vault and get the access token and revoke token from the vault
+    if (this.state.vault) {
+      // if the access token or revoke token is not available in the state, decrypt the vault and get the access token and revoke token from the vault
+      const { vaultData } = await this.#decryptAndParseVaultData({ password });
+      accessToken = accessToken || vaultData.accessToken;
+      revokeToken = revokeToken || vaultData.revokeToken;
+    }
+
+    // we should always throw an error if the access token or revoke token is not available
+    // to prevent the caller from using the controller in an invalid state
+
+    if (!accessToken) {
+      throw new Error(
+        SeedlessOnboardingControllerErrorMessage.InvalidAccessToken,
+      );
+    }
+
+    if (!revokeToken) {
+      throw new Error(
+        SeedlessOnboardingControllerErrorMessage.InvalidRevokeToken,
+      );
+    }
+
+    return { accessToken, revokeToken };
   }
 
   /**
@@ -1703,16 +1751,6 @@ export class SeedlessOnboardingController<
     }
   }
 
-  #assertIsSRPBackedUpUser(
-    value: unknown,
-  ): asserts value is SRPBackedUpUserDetails {
-    if (!this.state.authPubKey) {
-      throw new Error(
-        SeedlessOnboardingControllerErrorMessage.SRPNotBackedUpError,
-      );
-    }
-  }
-
   /**
    * Assert that the password is in sync with the global password.
    *
@@ -1778,7 +1816,7 @@ export class SeedlessOnboardingController<
    */
   async refreshAuthTokens(): Promise<void> {
     this.#assertIsAuthenticatedUser(this.state);
-    const { refreshToken, revokeToken } = this.state;
+    const { refreshToken } = this.state;
 
     const res = await this.#refreshJWTToken({
       connection: this.state.authConnection,
@@ -1793,6 +1831,7 @@ export class SeedlessOnboardingController<
     try {
       const { idTokens, accessToken, metadataAccessToken } = res;
       // re-authenticate with the new id tokens to set new node auth tokens
+      // NOTE: here we can't provide the `revokeToken` value to the `authenticate` method because `refreshAuthTokens` method can be called when the wallet (vault) is locked
       await this.authenticate({
         idTokens,
         accessToken,
@@ -1802,7 +1841,6 @@ export class SeedlessOnboardingController<
         groupedAuthConnectionId: this.state.groupedAuthConnectionId,
         userId: this.state.userId,
         refreshToken,
-        revokeToken,
         skipLock: true,
       });
     } catch (error) {
@@ -1834,11 +1872,6 @@ export class SeedlessOnboardingController<
         password,
         encryptionKey: vaultEncryptionKey,
       });
-      if (!revokeToken) {
-        throw new Error(
-          SeedlessOnboardingControllerErrorMessage.InvalidRevokeToken,
-        );
-      }
 
       const { newRevokeToken, newRefreshToken } = await this.#renewRefreshToken(
         {
@@ -2000,7 +2033,6 @@ export class SeedlessOnboardingController<
       const isNodeAuthTokenExpired = this.checkNodeAuthTokenExpired();
       const isMetadataAccessTokenExpired =
         this.checkMetadataAccessTokenExpired();
-
       // access token is only accessible when the vault is unlocked
       // so skip the check if the vault is locked
       let isAccessTokenExpired = false;
@@ -2088,6 +2120,9 @@ export class SeedlessOnboardingController<
     try {
       this.#assertIsAuthenticatedUser(this.state);
       const { accessToken } = this.state;
+      if (!accessToken) {
+        return true; // Consider missing token as expired
+      }
       const decodedToken = decodeJWTToken(accessToken);
       return decodedToken.exp < Math.floor(Date.now() / 1000);
     } catch {